データセーブ追加(NoiseBGMのみ)

+++ Assets/DataManage.cs
@@ -8,7 +8,7 @@ public class DataManage : MonoBehaviour
      * データの管理を担う。
      */
     //https://gametukurikata.com/program/savedata
-    public void SaveData(int highscore=-1,int score = -1, int operationMode = -1, int volumeSE = -1, int volumeNoise = -1, int volumeBGM = -1, int resolution = -1, int effect = -1, int weather = -1)
+    public void SaveData(int highscore=-1,int score = -1, int operationMode = -1, float volumeSE = -1, float volumeNoise = -1, float volumeBGM = -1, int resolution = -1, int effect = -1, int weather = -1)
      {
         /*
          * データをセーブする。
@@ -17,15 +17,15 @@ public class DataManage : MonoBehaviour
         if (highscore != -1) { PlayerPrefs.SetInt("highscore", highscore);}
         if (score != -1) { PlayerPrefs.SetInt("score", score);}
         if (operationMode != -1) { PlayerPrefs.SetInt("operationMode", operationMode);}
-        if (volumeSE != -1) { PlayerPrefs.SetInt("volumeSE", volumeSE);}
-        if (volumeNoise != -1) { PlayerPrefs.SetInt("volumeNoise", volumeNoise);}
-        if (volumeBGM != -1) { PlayerPrefs.SetInt("volumeBGM", volumeBGM);}
+        if (volumeSE != -1.0) { PlayerPrefs.SetFloat("volumeSE", volumeSE);}
+        if (volumeNoise != -1.0) { PlayerPrefs.SetFloat("volumeNoise", volumeNoise);}
+        if (volumeBGM != -1.0) { PlayerPrefs.SetFloat("volumeBGM", volumeBGM);}
         if (resolution != -1) { PlayerPrefs.SetInt("resolution", resolution);}
         if (effect != -1) { PlayerPrefs.SetInt("effect", effect);}
         if (weather != -1) { PlayerPrefs.SetInt("weather", weather);}
     }
 
-    public (int highscore,int score,int operationMode,int volumeSE,int volumeNoise,int VolumeBGM,int resolution,int effect,int weather) LoadData()
+    public (int highscore,int score,int operationMode,float volumeSE,float volumeNoise,float VolumeBGM,int resolution,int effect,int weather) LoadData()
     {
         /*
          * データをロードする
@@ -34,9 +34,9 @@ public class DataManage : MonoBehaviour
         int highscore = PlayerPrefs.GetInt("highscore", 0);
         int score = PlayerPrefs.GetInt("score", 0);
         int operationMode = PlayerPrefs.GetInt("operationMode", 1);
-        int volumeSE = PlayerPrefs.GetInt("volumeSE", 50);
-        int volumeNoise = PlayerPrefs.GetInt("volumeNoise", 50);
-        int volumeBGM = PlayerPrefs.GetInt("volumeBGM", 50);
+        float volumeSE = PlayerPrefs.GetFloat("volumeSE", 50);
+        float volumeNoise = PlayerPrefs.GetFloat("volumeNoise", 50);
+        float volumeBGM = PlayerPrefs.GetFloat("volumeBGM", 50);
         int resolution = PlayerPrefs.GetInt("resolution", 1);
         int effect = PlayerPrefs.GetInt("effect", 1);
         int weather = PlayerPrefs.GetInt("weather", 1);
